refactor(billing): expose public repository boundary
