+++ .dependency-cruiser-known-violations.json
@@ -8197,23 +8197,6 @@
   {
     "type": "dependency",
     "from": "src/app/routes/lazyPages.tsx",
-    "to": "src/features/daily/components/table/TableDailyRecordPage.tsx",
-    "unresolvedTo": "@/features/daily/components/table/TableDailyRecordPage",
-    "dependencyTypes": [
-      "aliased",
-      "aliased-tsconfig",
-      "aliased-tsconfig-paths",
-      "local",
-      "dynamic-import"
-    ],
-    "rule": {
-      "severity": "error",
-      "name": "no-external-feature-internals"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "src/app/routes/lazyPages.tsx",
     "to": "src/features/records/RecordList.tsx",
     "unresolvedTo": "@/features/records/RecordList",
     "dependencyTypes": [
@@ -9448,23 +9431,6 @@
     "from": "src/pages/AttendanceRecordPage.tsx",
     "to": "src/features/daily/components/pages/FullScreenDailyDialogPage.tsx",
     "unresolvedTo": "@/features/daily/components/pages/FullScreenDailyDialogPage",
-    "dependencyTypes": [
-      "aliased",
-      "aliased-tsconfig",
-      "aliased-tsconfig-paths",
-      "local",
-      "import"
-    ],
-    "rule": {
-      "severity": "error",
-      "name": "no-external-feature-internals"
-    }
-  },
-  {
-    "type": "dependency",
-    "from": "src/pages/BillingPage.tsx",
-    "to": "src/features/billing/hooks/useBillingSummary.ts",
-    "unresolvedTo": "@/features/billing/hooks/useBillingSummary",
     "dependencyTypes": [
       "aliased",
       "aliased-tsconfig",
