+++ src/features/billing/index.ts
@@ -1,0 +1,9 @@
+export { useBillingSummary } from './hooks/useBillingSummary';
+export { useBillingOrderRepository } from './hooks/useBillingOrderRepository';
+export { useBillingOrders, billingOrdersQueryKey } from './useBillingOrders';
+export type {
+  BillingOrderRepository,
+  BillingPersistenceDiagnostics,
+  BillingPersistenceDiagnosticStatus,
+} from './ports/billingOrderRepository';
+export type { BillingOrder } from './types';
